pyproject.toml
[project]
name = "llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index-embeddings-ollama>=0.8.1",
    "llama-index-llms-ollama>=0.7.1",
]
